Hoist IPC symmetric-registration machinery into the Backend base class
- Move IPC symmetric buffer registration (VMM alias create/teardown, IPC
  handle export/exchange, peer-mapping publication, and collective
  success/length consensus) from the IPC backend into the shared Backend
  base class so it can be reused by other backends.
- Reduce the IPC backend to delegate to register_ipc_symm_region /
  unregister_ipc_symm_region; behavior is unchanged.
